# manage-service/docs/swagger.yaml
definitions:
  entity.Applicant:
    properties:
      aggregated_score:
        type: number
      created_at:
        type: string
      evaluation_status:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      patronymic:
        type: string
      program_id:
        type: integer
      score:
        description: Maps to aggregated_score for legacy compatibility
        type: number
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.createApplicantRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      patronymic:
        type: string
      program_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing applicants and documents
  title: Manage Service API
  version: "1.0"
paths:
  /v1/applicants:
    get:
      description: Возвращает список абитуриентов. Можно фильтровать по ID программы.
      parameters:
      - description: ID образовательной программы для фильтрации
        in: query
        name: program_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Applicant'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список абитуриентов
      tags:
      - applicants
    post:
      consumes:
      - application/json
      description: Создаёт нового абитуриента и привязывает его к образовательной
        программе.
      parameters:
      - description: Данные нового абитуриента
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createApplicantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Applicant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание абитуриента
      tags:
      - applicants
  /v1/applicants/{id}:
    delete:
      description: Удаляет абитуриента и все связанные с ним данные по ID.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление абитуриента
      tags:
      - applicants
...
  /v1/applicants/{id}/data:
    get:
      description: Возвращает структурированные данные абитуриента по указанной категории
        (identification, education, transcript и т.д.).
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Категория данных
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Данные абитуриента по категории
      tags:
      - applicants
    patch:
      consumes:
      - application/json
      description: Обновляет данные абитуриента по указанной категории. Принимает
        произвольный JSON-объект.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Категория данных
        in: query
        name: category
        type: string
      - description: Обновлённые данные
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление данных абитуриента
      tags:
      - applicants
  /v1/applicants/{id}/data/{category}/{dataId}:
    delete:
      description: Удаляет конкретную запись данных абитуриента (например, запись
        о работе или достижении) по категории и ID записи.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Категория данных (work_experience, achievements и т.д.)
        in: path
        name: category
        required: true
        type: string
      - description: ID записи
        in: path
        name: dataId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление данных абитуриента
      tags:
      - applicants
...
  /v1/applicants/{id}/transfer-to-experts:
    post:
      description: Переводит абитуриента на этап экспертной оценки (изменяет статус
        заявки).
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Передача абитуриента экспертам
      tags:
      - applicants
  /v1/applicants/{id}/transfer-to-operator:
    post:
      description: Переводит абитуриента на этап проверки оператором (изменяет статус
        заявки).
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Передача абитуриента оператору
      tags:
      - applicants
  /v1/applicants/{id}/ws:
    get:
      description: Устанавливает WebSocket-соединение для получения обновлений статуса
        обработки документов абитуриента в реальном времени.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: WebSocket соединение
      tags:
      - applicants
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"